[Feature/#20] 홈화면, 글쓰기 구현
