+++ src/pages/WritePost/WritePost.tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import { css } from "@emotion/react";
 import {
   container,
   contentTitle,
